Move scorecard code to dedicated file
